+++ script/provider/provider.lua
@@ -61,6 +61,12 @@ m.attributes = {}
 function m.register(method)
     return function (attrs)
         m.attributes[method] = attrs
+        if attrs.preview and not PREVIEW then
+            return
+        end
+        if attrs.capability then
+            cap.filling(attrs.capability)
+        end
         proto.on(method, attrs[1])
     end
 end
@@ -101,7 +107,7 @@ m.register 'initialize' {
         end
 
         return {
-            capabilities = cap.getIniter(),
+            capabilities = cap.getProvider(),
             serverInfo   = {
                 name    = 'sumneko.lua',
             },
@@ -261,6 +267,9 @@ m.register 'textDocument/didChange' {
 }
 
 m.register 'textDocument/hover' {
+    capability = {
+        hoverProvider = true,
+    },
     abortByFileUpdate = true,
     ---@async
     function (params)
@@ -299,6 +308,9 @@ m.register 'textDocument/hover' {
 }
 
 m.register 'textDocument/definition' {
+    capability = {
+        definitionProvider = true,
+    },
     abortByFileUpdate = true,
     ---@async
     function (params)
@@ -338,6 +350,9 @@ m.register 'textDocument/definition' {
 }
 
 m.register 'textDocument/typeDefinition' {
+    capability = {
+        typeDefinitionProvider = true,
+    },
     abortByFileUpdate = true,
     ---@async
     function (params)
@@ -377,6 +392,9 @@ m.register 'textDocument/typeDefinition' {
 }
 
 m.register 'textDocument/references' {
+    capability = {
+        referencesProvider = true,
+    },
     abortByFileUpdate = true,
     ---@async
     function (params)
@@ -404,10 +422,14 @@ m.register 'textDocument/references' {
 }
 
 m.register 'textDocument/documentHighlight' {
-    abortByFileUpdate = true,
+    capability = {
+        documentHighlightProvider = true,
+    },
+    ---@async
     function (params)
         local core = require 'core.highlight'
         local uri  = files.getRealUri(params.textDocument.uri)
+        workspace.awaitReady(uri)
         if not files.exists(uri) then
             return nil
         end
@@ -428,6 +450,11 @@ m.register 'textDocument/documentHighlight' {
 }
 
 m.register 'textDocument/rename' {
+    capability = {
+        renameProvider = {
+            prepareProvider = true,
+        },
+    },
     abortByFileUpdate = true,
     ---@async
     function (params)
@@ -631,6 +658,11 @@ m.register 'completionItem/resolve' {
 }
 
 m.register 'textDocument/signatureHelp' {
+    capability = {
+        signatureHelpProvider = {
+            triggerCharacters = { '(', ',' },
+        },
+    },
     abortByFileUpdate = true,
     ---@async
     function (params)
@@ -677,6 +709,9 @@ m.register 'textDocument/signatureHelp' {
 }
 
 m.register 'textDocument/documentSymbol' {
+    capability = {
+        documentSymbolProvider = true,
+    },
     abortByFileUpdate = true,
     ---@async
     function (params)
@@ -723,6 +758,17 @@ m.register 'textDocument/documentSymbol' {
 }
 
 m.register 'textDocument/codeAction' {
+    capability = {
+        codeActionProvider = {
+            codeActionKinds = {
+                '',
+                'quickfix',
+                'refactor.rewrite',
+                'refactor.extract',
+            },
+            resolveProvider = false,
+        },
+    },
     abortByFileUpdate = true,
     function (params)
         local core        = require 'core.code-action'
@@ -757,6 +803,17 @@ m.register 'textDocument/codeAction' {
 }
 
 m.register 'workspace/executeCommand' {
+    capability = {
+        executeCommandProvider = {
+            commands = {
+                'lua.removeSpace',
+                'lua.solve',
+                'lua.jsonToLua',
+                'lua.setConfig',
+                'lua.autoRequire',
+            },
+        },
+    },
     ---@async
     function (params)
         local command = params.command:gsub(':.+', '')
@@ -780,6 +837,9 @@ m.register 'workspace/executeCommand' {
 }
 
 m.register 'workspace/symbol' {
+    capability = {
+        workspaceSymbolProvider = true,
+    },
     abortByFileUpdate = true,
     ---@async
     function (params)
@@ -826,7 +886,29 @@ m.register 'textDocument/semanticTokens/full' {
     end
 }
 
+local function toArray(map)
+    local array = {}
+    for k in pairs(map) do
+        array[#array+1] = k
+    end
+    table.sort(array, function (a, b)
+        return map[a] < map[b]
+    end)
+    return array
+end
+
+
 m.register 'textDocument/semanticTokens/range' {
+    capability = {
+        semanticTokensProvider = {
+            legend = {
+                tokenTypes     = toArray(define.TokenTypes),
+                tokenModifiers = toArray(define.TokenModifiers),
+            },
+            range = true,
+            full  = false,
+        },
+    },
     ---@async
     function (params)
         log.debug('textDocument/semanticTokens/range')
@@ -843,11 +925,14 @@ m.register 'textDocument/semanticTokens/range' {
 }
 
 m.register 'textDocument/foldingRange' {
-    abortByFileUpdate = true,
+    capability = {
+        foldingRangeProvider = true,
+    },
     ---@async
     function (params)
         local core    = require 'core.folding'
         local uri     = files.getRealUri(params.textDocument.uri)
+        workspace.awaitReady(uri)
         if not files.exists(uri) then
             return nil
         end
@@ -915,6 +1000,10 @@ m.register '$/status/click' {
 }
 
 m.register 'textDocument/formatting' {
+    capability = {
+        documentFormattingProvider = true,
+    },
+    preview = true,
     ---@async
     function(params)
         local uri = files.getRealUri(params.textDocument.uri)
@@ -923,11 +1012,15 @@ m.register 'textDocument/formatting' {
             return nil
         end
 
+        if not config.get(uri, 'Lua.format.enable') then
+            return nil
+        end
+
         local pformatting = require 'provider.formatting'
         pformatting.updateConfig(uri)
 
         local core = require 'core.formatting'
-        local edits = core(uri)
+        local edits = core(uri, params.options)
         if not edits or #edits == 0 then
             return nil
         end
@@ -945,6 +1038,10 @@ m.register 'textDocument/formatting' {
 }
 
 m.register 'textDocument/rangeFormatting' {
+    capability = {
+        documentRangeFormattingProvider = true,
+    },
+    preview = true,
     ---@async
     function(params)
         local uri = files.getRealUri(params.textDocument.uri)
@@ -953,11 +1050,15 @@ m.register 'textDocument/rangeFormatting' {
             return nil
         end
 
+        if not config.get(uri, 'Lua.format.enable') then
+            return nil
+        end
+
         local pformatting = require 'provider.formatting'
         pformatting.updateConfig(uri)
 
         local core = require 'core.rangeformatting'
-        local edits = core(uri, params.range)
+        local edits = core(uri, params.range, params.options)
         if not edits or #edits == 0 then
             return nil
         end
@@ -975,6 +1076,12 @@ m.register 'textDocument/rangeFormatting' {
 }
 
 m.register 'textDocument/onTypeFormatting' {
+    capability = {
+        documentOnTypeFormattingProvider = {
+            firstTriggerCharacter = '\n',
+            moreTriggerCharacter  = nil, -- string[]
+        },
+    },
     abortByFileUpdate = true,
     ---@async
     function (params)